employee_management/pyproject.toml
[project]
name = "employee-management"
version = "0.1.0"
description = ""
authors = [
    {name = "anjuclaysys",email = "anju.vijayan@claysys.net"}
]
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.131.0,<0.132.0)",
    "sqlalchemy (>=2.0.46,<3.0.0)",
    "uvicorn (>=0.41.0,<0.42.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "pyjwt (>=2.11.0,<3.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "bcrypt (==3.2.2)",
    "python-json-logger (>=4.0.0,<5.0.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "email-validator (>=2.3.0,<3.0.0)"
]

[tool.poetry]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
